SEChecker,LibSEFS:    converted basic infrastructure to qpol    modified the fc_entry free function to be suitable to pass to vector destroy Bugz: 1781

+++ sechecker/register_list.h
@@ -42,8 +42,8 @@ extern int unreachable_doms_register(sechk_lib_t *lib);
 extern int imp_range_trans_register(sechk_lib_t *lib);
 /* TODO: additional externs go here ... */
 
-int sechk_register_list_get_num_modules();
+size_t sechk_register_list_get_num_modules();
 const sechk_module_name_reg_t* sechk_register_list_get_modules();
-int sechk_register_list_get_num_profiles();
+size_t sechk_register_list_get_num_profiles();
 const sechk_profile_name_reg_t* sechk_register_list_get_profiles();
 #endif /* SECHK_REGISTER_LIST_H */
